[VDS-250] adding mermaidGraph component
