+++ src/components/Panel.tsx
@@ -2,12 +2,14 @@ import React, { useEffect, useRef, useState } from "react";
 import { API, useParameter } from "@storybook/manager-api";
 import { AddonPanel, P } from "@storybook/components";
 import { getGraphRender } from "../scripts/utils";
-import { PARAM_KEY } from "../scripts/constants";
+import { GRAPH_PARAM_KEY } from "../scripts/constants";
 import { DepGraphParams } from "../scripts/types";
 import { depGraphDefaultParams } from "../scripts/data";
 
 //TODO this section needs to be isolated with the UI render component
 import mermaid from "mermaid";
+import MermaidGraph from "./MermaidGraph";
+
 mermaid.initialize({ startOnLoad: false });
 
 interface PanelProps {
@@ -16,26 +18,11 @@ interface PanelProps {
 }
 
 function Panel({ active, api }: PanelProps) {
-  const currentStoryData = api.getCurrentStoryData();
-  const params = useParameter<Partial<DepGraphParams>>(
-    PARAM_KEY,
+  const dependencyData = useParameter<Partial<DepGraphParams>>(
+    GRAPH_PARAM_KEY,
     depGraphDefaultParams,
   );
-  const normalizedParams = Object.assign(params, depGraphDefaultParams);
-  const render = getGraphRender(normalizedParams, currentStoryData);
-
-  const mermaidRef = useRef(null);
-
-  useEffect(() => {
-    //TODO this section needs to be isolated with the UI render component
-    mermaid
-      .run()
-      .then(() => {
-        mermaidRef.current &&
-          mermaidRef.current.removeAttribute("data-processed");
-      })
-      .catch(() => console.error("Mermaid Render Failed"));
-  }, [render.value]);
+  const currentStoryData = api.getCurrentStoryData();
 
   return (
     <AddonPanel active={active}>
@@ -44,16 +31,10 @@ function Panel({ active, api }: PanelProps) {
         <h2> Story ID: {currentStoryData.id}</h2>
         <h2> Story Import Path: {currentStoryData.importPath}</h2>
         {/*TODO Create a component for graph displaying and navigation. Consider conditional components based on UI_Type*/}
-        <div style={{ background: "white", color: "black", padding: "16px" }}>
-          <h3> Graph: </h3>
-          {render.error ? (
-            <p> {render.value} </p>
-          ) : (
-            <p className={"mermaid"} ref={mermaidRef}>
-              {render.value}
-            </p>
-          )}
-        </div>
+        <MermaidGraph
+          currentStoryData={currentStoryData}
+          dependencyData={dependencyData}
+        ></MermaidGraph>
       </div>
     </AddonPanel>
   );
