+++ src/scripts/constants.ts
@@ -2,7 +2,7 @@ import { name as packageName } from "../../package.json"
 export const ADDON_ID = packageName;
 export const PANEL_ID = `${ADDON_ID}/panel`;
 export const PANEL_NAME = `DepGraph`;
-export const PARAM_KEY = `dependencyGraph`;
+export const GRAPH_PARAM_KEY = `dependencyGraph`;
 export const uiTypes = {
   mermaid: 'Mermaid'
 } as const
